refactor: pull out send_and_confirm_tx into utils file
