+++ src/lib.rs
@@ -15,4 +15,5 @@ pub mod snapshot;
 pub mod spinner;
 pub mod update_metadata;
 pub mod uses;
+pub mod utils;
 pub mod withdraw;
