+++ src/uses.rs
@@ -5,42 +5,14 @@ use mpl_token_metadata::{
     id as metadata_program_id,
     instruction::{approve_use_authority, revoke_use_authority, utilize},
 };
-use retry::{delay::Exponential, retry};
 use solana_client::rpc_client::RpcClient;
-use solana_sdk::{
-    instruction::Instruction, pubkey::Pubkey, signature::Keypair, signer::Signer,
-    transaction::Transaction,
-};
+use solana_sdk::{pubkey::Pubkey, signer::Signer};
 
 use crate::{
     derive::{derive_metadata_pda, derive_use_authority_record},
     parse::{parse_keypair, parse_solana_config},
+    utils::send_and_confirm_transaction,
 };
-
-fn send_and_confirm_transaction(
-    client: &RpcClient,
-    keypair: Keypair,
-    instructions: &[Instruction],
-) -> Result<String> {
-    let recent_blockhash = client.get_latest_blockhash()?;
-    let tx = Transaction::new_signed_with_payer(
-        instructions,
-        Some(&keypair.pubkey()),
-        &[&keypair],
-        recent_blockhash,
-    );
-
-    // Send tx with retries.
-    let res = retry(
-        Exponential::from_millis_with_factor(250, 2.0).take(3),
-        || client.send_and_confirm_transaction(&tx),
-    );
-
-    let sig = res?;
-
-    println!("TxId: {}", sig);
-    Ok(sig.to_string())
-}
 
 pub fn approve_use_delegate(
     client: &RpcClient,
