+++ src/utils.rs
@@ -1,0 +1,31 @@
+use anyhow::Result;
+use retry::{delay::Exponential, retry};
+use solana_client::rpc_client::RpcClient;
+use solana_sdk::{
+    instruction::Instruction, signature::Keypair, signer::Signer, transaction::Transaction,
+};
+
+pub fn send_and_confirm_transaction(
+    client: &RpcClient,
+    keypair: Keypair,
+    instructions: &[Instruction],
+) -> Result<String> {
+    let recent_blockhash = client.get_latest_blockhash()?;
+    let tx = Transaction::new_signed_with_payer(
+        instructions,
+        Some(&keypair.pubkey()),
+        &[&keypair],
+        recent_blockhash,
+    );
+
+    // Send tx with retries.
+    let res = retry(
+        Exponential::from_millis_with_factor(250, 2.0).take(3),
+        || client.send_and_confirm_transaction(&tx),
+    );
+
+    let sig = res?;
+
+    println!("TxId: {}", sig);
+    Ok(sig.to_string())
+}
